Patient Graph is now a component
