+++ .eslintrc.json
@@ -4,6 +4,7 @@
     "jest": true
   },
   "rules": {
-    "linebreak-style": "off"
+    "linebreak-style": "off",
+    "no-console": "off"
   }
 }
